+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Changelog ##
 #### dev ####
+* Add localization support for MTS strings
 
 #### v3.30.3 ####
 * Fix package information being lost for patched classes (Alchyr)